+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed calling `PybricksHub.write()` methods.
 
+### Changed
+- Downloading programs without starting them can now be done by
+  adding `--no-start` to the run tool.
+
 ### Removed
 - Removed `REPLHub`. This was used for non-Pybricks MicroPython boards, but
   `mpremote` should be used for this.